requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
openai==0.27.0
requests==2.28.1
python-dotenv==0.19.0
gunicorn==20.1.0